had to resolve some conflicts in Main file, moved objectCache object into main com.revature.objectmapper package - dont use name based packages any more like jack, aron, etc.

+++ src/main/java/com/revature/app/ProducerCountry.java
@@ -56,7 +56,7 @@ public class ProducerCountry {
 		this.numManufactuers = numManufactuers;
 	}
 
-	public int getTotalAnnual() {
+	public long getTotalAnnual() {
 		return totalAnnual;
 	}
